% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_box.R
\name{bio_box}
\alias{bio_box}
\title{bio_box}
\usage{
bio_box(matrix)
}
\arguments{
\item{matrix}{numeric}
}
\value{
plot
}
\description{
A partir de una matriz de datos, realiza grafico de cajas y bigotes.
}
\examples{
set.seed(0)
 matrix <- round(matrix(runif(30, min = 0, max = 100), nrow = 10, ncol = 3), 2)
 bio_box(matrix)
}
