% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMetrix.R
\name{bio_mean}
\alias{bio_mean}
\title{BioMetrix}
\usage{
bio_mean(matrix)
}
\arguments{
\item{matrix, }{numeric}
}
\value{
matrix, mean
}
\description{
A partir de una matriz de datos, calcula la media.
}
\examples{
matrix <- matrix(1:9, nrow = 3, ncol = 3)
bio_mean(matrix)

}
