% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_sd.R
\name{bio_sd}
\alias{bio_sd}
\title{bio_sd}
\usage{
bio_sd(matrix)
}
\arguments{
\item{matrix, }{numeric}
}
\value{
matrix, sd
}
\description{
A partir de una matriz de datos, calcula la desviacion estandar
}
\examples{
 set.seed(0)
 matrix <- round(matrix(runif(30, min = 0, max = 100), nrow = 10, ncol = 3), 2)
bio_sd(matrix)
}
