% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_mean.R
\name{bio_mean}
\alias{bio_mean}
\title{bio_mean}
\usage{
bio_mean(matrix)
}
\arguments{
\item{matrix, }{numeric}
}
\value{
matrix, mean
}
\description{
A partir de una matriz de datos, calcula la media por columna.
}
\examples{
set.seed(0)
 matrix <- round(matrix(runif(30, min = 0, max = 100), nrow = 10, ncol = 3), 2)
bio_mean(matrix)

}
